+++ CHANGELOG.md
@@ -14,6 +14,10 @@ any parts of the framework not mentioned in the documentation should generally b
 
 * Added support for Python 3.11.
 
+### Changed
+
+* Added support to overwrite serializer methods in customized schema class
+
 ## [6.0.0] - 2022-09-24
 
 ### Fixed